feat(Switch Node): Add a new version of Switch node with dynamic outputs
