+++ packages/nodes-base/nodes/Switch/V1/test/switch.node.test.ts
