+++ packages/nodes-base/nodes/Switch/V1/test/switch.rules.workflow.json
